renames helper modules, adds margin to create fika button
